+++ source/blender/blenloader/BLO_readfile.h
@@ -19,6 +19,8 @@
 #ifndef __BLO_READFILE_H__
 #define __BLO_READFILE_H__
 
+#include "BLI_sys_types.h"
+
 /** \file
  * \ingroup blenloader
  * \brief external readfile function prototypes.
@@ -145,9 +147,7 @@ void BLO_library_link_end(struct Main *mainl,
                           struct ViewLayer *view_layer,
                           const struct View3D *v3d);
 
-int BLO_library_link_copypaste(struct Main *mainl,
-                               BlendHandle *bh,
-                               const unsigned int id_types_mask);
+int BLO_library_link_copypaste(struct Main *mainl, BlendHandle *bh, const uint64_t id_types_mask);
 
 void *BLO_library_read_struct(struct FileData *fd, struct BHead *bh, const char *blockname);
